+++ system/event-script-engine/src/test/java/com/accenture/tasks/HeaderParserFunction.java
@@ -20,22 +20,14 @@ package com.accenture.tasks;
 
 import org.platformlambda.core.annotations.PreLoad;
 import org.platformlambda.core.models.TypedLambdaFunction;
-import org.platformlambda.core.util.Utility;
 
 import java.util.Map;
 
-@PreLoad(route="numeric.decision", instances=10)
-public class NumericDecision implements TypedLambdaFunction<Map<String, Object>, Integer> {
-
-    private static final String DECISION = "decision";
+@PreLoad(route="header.parser", instances=10)
+public class HeaderParserFunction implements TypedLambdaFunction<Map<String, Object>, Map<String, Object>> {
 
     @Override
-    public Integer handleEvent(Map<String, String> headers, Map<String, Object> input, int instance) {
-        if (input.containsKey(DECISION)) {
-            return Utility.getInstance().str2int(input.get(DECISION).toString());
-
-        } else {
-            throw new IllegalArgumentException("Missing decision");
-        }
+    public Map<String, Object> handleEvent(Map<String, String> headers, Map<String, Object> input, int instance) throws Exception {
+        return input;
     }
 }